% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{block}
\alias{block}
\title{block}
\usage{
block(dfA, dfB, cols.exact, cols.numeric, cols.encoder, 
encoder.trainA, encoder.trainB,  encoder.block.method = 'binary', numeric.range, 
model.path, dim.latent, dim.encode, dim.decode, max.length,
num.encode.layers, num.decode.layers, batch.size, epochs, lr,
validation.split, save.dir, earlystop, earlystop.patience, 
tensorboard, tensorboard.runid)
}
\arguments{
\item{dfA}{Dataframe to be linked to \code{dfB}}

\item{dfB}{Dataframe to be linked to \code{dfA}, if doing deduplication this is \code{dfA}}

\item{cols.exact}{List of exact match columns}

\item{cols.numeric}{List of numeric range columns}

\item{numeric.range}{Range of numeric}

\item{cols.encoder}{List of encoder columns}

\item{encoder.model.path}{Path to encoder model}

\item{encoder.trainA}{Vector of names to train}

\item{encoder.trainB}{Vector of matching names to train}

\item{encoder.block.method}{binary or cluster}

\item{encoder.nclusters}{Number of cluster is encoding by cluster}

\item{known.matches}{Dataframe of known matches with first column having indices of matches 
from dfA and second column having indices of known matches from dfB}

\item{dim.latent}{Number of latent dimensions}

\item{dim.encode}{Number of encoding dimensions}

\item{dim.decode}{Number of decoding dimensions}

\item{max.length}{Maximum length of characters}
}
\value{
Nested list of blocks where each block is one encoding. 
Sub-lists are of indices of namesA and namesB belonging to each block
}
\description{
Block data to reduce the number of comparisons needed for linkage
}
