% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{blockClusterEncoder}
\alias{blockClusterEncoder}
\title{blockCluster}
\usage{
blockClusterEncoder(encoded.A, encoded.B, n.cores, n.centers = 20,
  iter.max = 1000)
}
\arguments{
\item{encoded.A}{List of matrices from encoded columns in dfA}

\item{encoded.B}{List of matrices from encoded columns in dfB}

\item{n.centers}{Number of clusters to use}

\item{iter.max}{Maximum number of iterations}
}
\description{
blockCluster
}
