% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{blockExact}
\alias{blockExact}
\title{blockExact}
\usage{
blockExact(dfA, dfB, cols.exact, n.cores)
}
\arguments{
\item{dfA}{Dataframe to be linked to `dfB`}

\item{dfB}{Dataframe to be linked to `dfA`}

\item{cols.exact}{List with columns in dfA to be exactly matched to columns in dfB}

\item{n.cores}{cores}
}
\value{
Nested list where each list corresponds to one column in cols.exact. Each sublist 
contains a vector of indices named `A` and a vector of indices named `B` for indices in `dfA`
and `dfB` which correspond to the same block value
}
\description{
Block on exact column values
}
