% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockMetrics.R
\name{blockMetrics}
\alias{blockMetrics}
\title{blockMetrics}
\usage{
blockMetrics(dfA, dfB, blocks, known.matches = NULL)
}
\arguments{
\item{dfA}{dataframe A}

\item{dfB}{dataframe B}

\item{blocks}{Blocks}

\item{known.matches}{indices}
}
\value{
List of metrics
}
\description{
Reports block metrics to assess quality of blocking. If known matches are null then
}
