% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{blockNumeric}
\alias{blockNumeric}
\title{blockNumeric}
\usage{
blockNumeric(dfA, dfB, n.cores, cols.numeric, numeric.range = 1)
}
\arguments{
\item{dfA}{Dataframe to be linked to `dfB`}

\item{dfB}{Dataframe to be linked to `dfA`}

\item{cols.numeric}{Named list}

\item{numeric.range}{Window}
}
\description{
blockNumeric
}
