% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainModel.R
\docType{data}
\name{CheckReconstruction}
\alias{CheckReconstruction}
\title{CheckReconstruction}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CheckReconstruction
}
\description{
Custom callback to check reconstruction during training
}
\keyword{datasets}
