% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{compare}
\usage{
compare(dfA, dfB, blocks, compare.string.encoder = NULL,
  encoder.model.path = NULL, compare.string.sim = NULL,
  string.sim.method = "jw", compare.numeric = NULL, compare.exact = NULL,
  n.cores = detectCores() - 1)
}
\value{
List
}
\description{
Generate comparison matrix for pairs of records
}
