% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedLetters.R
\name{disembedLetters}
\alias{disembedLetters}
\title{disembedLetters}
\usage{
disembedLetters(embeddedArray)
}
\arguments{
\item{embeddedArray}{3-dim array of one-hot encoded characters}
}
\value{
Character vector with length dim(embeddedArray)[1]
}
\description{
Revert embedded array back into a character vector
}
