% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedLetters.R
\name{disembedName}
\alias{disembedName}
\title{disembedName}
\usage{
disembedName(x)
}
\arguments{
\item{x}{integer vector}
}
\value{
character vector with one name
}
\description{
Disembed a single name
}
