% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedLetters.R
\name{embedLetters}
\alias{embedLetters}
\title{embedLetters}
\usage{
embedLetters(names, max.length = 12)
}
\arguments{
\item{names}{Character vector}

\item{max.length}{Numeric representing number of letters to keep for each name. 
Names shorter than \code{max.length} will be padded with 0s from the right}
}
\value{
Binary array with dimension: length(names) x max.length x 28
}
\description{
Embed a character vector into one-hot encoded array
}
\details{
One-hot encoding will have dimension 28 for 27 lower case letters as well as a space
}
\examples{
\dontrun{
# Sample names 
names <- c('amy', 'john', 'cathy')

# Embed 
embedded <- embedLetters(names, max.length = 12)

dim(embedded)
> [1] 3  12  28
}

}
