% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{encodeCompare}
\alias{encodeCompare}
\title{encodeCompare}
\usage{
encodeCompare(vec.A, vec.B, encoder.model.path = NULL)
}
\arguments{
\item{vec.A}{exact vector from dfA}

\item{vec.B}{exact vector from dfB}

\item{encoder.model.path}{model path}
}
\value{
vector comparisons
}
\description{
encodeCompare
}
