% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{encoder_iowa_first_4}
\alias{encoder_iowa_first_4}
\title{Encoder Model Iowa First Names 4 dim}
\format{RAW}
\usage{
encoder_iowa_first_4
}
\description{
Encoder Model Iowa First Names 4 dim
}
\keyword{Keras}
\keyword{encoder}
\keyword{model}
