% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{Modle}
\name{encoder_iowa_first_512}
\alias{encoder_iowa_first_512}
\title{Iowa First Name 512-dim Encoder}
\format{Keras model}
\usage{
encoder_iowa_first_512(object)
}
\description{
Iowa First Name 512-dim Encoder
}
