% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{encodeVector}
\alias{encodeVector}
\title{encodeVector}
\usage{
encodeVector(vec, model)
}
\arguments{
\item{vec}{Character vector to encoded}

\item{model}{Keras encoding model or path to Keras h5 model}
}
\value{
Matrix of encoded values with dimensions length(vec) x dim.latent
}
\description{
Encodes a character vector with existing model
}
