% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{exactCompare}
\alias{exactCompare}
\title{exactCompare}
\usage{
exactCompare(vec.A, vec.B)
}
\arguments{
\item{vec.A}{exact vector from dfA}

\item{vec.B}{exact vector from dfB}
}
\value{
vector 1 for exact match, 0 for no match
}
\description{
exactCompare
}
