% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{getIndex}
\alias{getIndex}
\title{getIndex}
\usage{
getIndex(x, vector)
}
\arguments{
\item{x}{Vector of values to be matched}

\item{vector}{Vector to get indices from}
}
\value{
Indicies of vector which match values in x
}
\description{
getIndex
}
