% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iowa_sample_matches}
\alias{iowa_sample_matches}
\title{Iowa Census Sample Matches}
\format{Dataframe with 4320 observations}
\usage{
iowa_sample_matches
}
\description{
Iowa Census Sample Matches
}
\keyword{datasets}
