% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iowa_true_matches}
\alias{iowa_true_matches}
\title{Iowa Census True Matches}
\format{Dataframe with 1000 observations}
\usage{
iowa_true_matches
}
\description{
Iowa Census True Matches
}
\keyword{datasets}
