% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadEncoders.R
\name{loadSampleEncoders}
\alias{loadSampleEncoders}
\title{loadSampleEncoders}
\usage{
loadSampleEncoders()
}
\value{
Example encoder models for Iowa dataset
}
\description{
Loads sample encoder models trained on Iowa dataset first names
}
