% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{numCompare}
\alias{numCompare}
\title{numCompare}
\usage{
numCompare(vec.A, vec.B)
}
\arguments{
\item{vec.A}{exact vector from dfA}

\item{vec.B}{exact vector from dfB}
}
\value{
vector comparisons
}
\description{
numCompare
}
