% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{stringCompare}
\alias{stringCompare}
\title{stringCompare}
\usage{
stringCompare(vec.A, vec.B, method = "jw")
}
\arguments{
\item{vec.A}{exact vector from dfA}

\item{vec.B}{exact vector from dfB}

\item{method}{Jaro Winkler}
}
\value{
dataframe difference
}
\description{
stringCompare
}
