% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{blockBinaryEncoder}
\alias{blockBinaryEncoder}
\title{blockBinaryEncoder}
\usage{
blockBinaryEncoder(encoded.A, encoded.B, n.cores)
}
\arguments{
\item{encoded.A}{Matrix of encoded values}

\item{encoded.B}{Matrix of encoded values}

\item{n.cores}{Number of cores to parallelize}
}
\value{
List
}
\description{
blockBinaryEncoder
}
