% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockMetrics.R
\name{blockMetrics}
\alias{blockMetrics}
\title{blockMetrics}
\usage{
blockMetrics(dfA, dfB, blocks, known.matches = NULL)
}
\value{
List of metrics
}
\description{
Reports block metrics to assess quality of blocking. If known matches are null then
}
