% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanNames.R
\name{cleanNames}
\alias{cleanNames}
\title{cleanNames}
\usage{
cleanNames(df, cols, split = FALSE, split.pattern = ",",
  new.col.names = NULL, lower = TRUE, strip.punctuation = TRUE)
}
\arguments{
\item{df}{Dataframe}

\item{cols}{Vector of column names to clean}

\item{split}{if at least one of the columns in cols needs to be split}

\item{split.pattern}{Vector of pattern to split on in order of columns to be split}

\item{new.col.names}{Named list. Names are names of columns to be split, values are 
vectors of new column names post split.}

\item{lower}{cols are lower-cased, if FALSE cols are returned with case as is}

\item{strip.punctuation}{punctuation is stripped from cols, if FALSE all 
puntuation is left as is}
}
\value{
Dataframe with processed character columns
}
\description{
cleanNames is used to preprocess character columns for blocking and linking functions.
Names can be standardized by converting all characters to lower case, splitting full name 
columns into individual parts such as 'first' and 'last', and punctuation can be removed. 
For blocking and linking on autoencoded name fields, it is recommended to split each part of 
the name such as first, middle, and last, into separate columns.
}
