% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{encode}
\alias{encode}
\title{encode}
\usage{
encode(dfA, dfB, cols.encoder, encoder.model.path = NULL,
  encoder.trainA = NULL, encoder.trainB = NULL, dim.latent = 8,
  dim.encode = 64, dim.decode = 64, max.length = 12,
  num.encode.layers = 2, num.decode.layers = 2, batch.size = 32,
  epochs = 500, lr = 5e-04, validation.split = 0.2,
  save.dir = "~/blocking_models/", reconstruct = TRUE, reconstruct.n = 5,
  reconstruct.display = 20, earlystop = FALSE, earlystop.patience = 10,
  tensorboard = FALSE, tensorboard.runid = as.character(Sys.time()),
  verbose = 2)
}
\description{
Adds encoded columns to dataframes to be linked
}
