% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{Model}
\name{encoder_iowa_first_4}
\alias{encoder_iowa_first_4}
\title{Iowa First Name 4-dim Encoder}
\format{Keras model}
\usage{
encoder_iowa_first_4(object)
}
\description{
Iowa First Name 4-dim Encoder
}
