% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iowa_1915}
\alias{iowa_1915}
\title{Sample Iowa Census 1915}
\format{Dataframe containing 6881 observations.
\describe{
\item{name}
\item{birthyear}
}}
\usage{
iowa_1915
}
\description{
Subset of Iowa Census 1915
}
\keyword{datasets}
