% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNames.R
\name{plotNames}
\alias{plotNames}
\title{plotNames}
\usage{
plotNames(names, n.samples = 1000, encoder.model.path = NULL,
  title = "Sample of Names in Latent Space Plotted with t-SNE",
  text.size = 3, seed = 0)
}
\description{
plotNames
}
