% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervisedLink.R
\name{supervisedLink}
\alias{supervisedLink}
\title{link}
\usage{
supervisedLink(dfA, dfB, train.indices, validation.split = 0.2,
  method = "log", blocks = NULL, block.exact = NULL,
  block.numeric = NULL, block.numeric.range = NULL, block.encoder = NULL,
  block.encoder.model.path = NULL, block.encoder.method = "cluster",
  comparisons = NULL, compare.string.encoder = NULL,
  compare.encoder.model.path = NULL, compare.string.sim = NULL,
  string.sim.method = "jw", compare.numeric = NULL, compare.exact = NULL,
  n.cores = parallel::detectCores() - 1)
}
\arguments{
\item{dfA}{dataframe}

\item{dfB}{dataframe}
}
\value{
list
}
\description{
Supervised record linkage
}
