% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainModel.R
\name{trainModel}
\alias{trainModel}
\title{trainModel}
\usage{
trainModel(namesA, namesB, dim.latent = 8, dim.encode = 64,
  dim.decode = 64, max.length = 12, num.encode.layers = 2,
  num.decode.layers = 2, batch.size = 32, epochs = 500, lr = 1e-04,
  validation.split = 0.2, save.dir = "~/blocking_models/",
  reconstruct = TRUE, reconstruct.n = 5, reconstruct.display = 20,
  earlystop = TRUE, earlystop.patience = 10, tensorboard = TRUE,
  tensorboard.runid = as.character(Sys.time()), verbose = 2)
}
\arguments{
\item{namesA}{Character vector of names}

\item{namesB}{Character vector of names known to be matches of namesA. 
Length of namesA and namesB must be equal.}

\item{dim.latent}{Number of dimensions to use in the latent vector}

\item{dim.encode}{Number of hidden units to use in each of the encoding layers}

\item{dim.decode}{Number of hidden units to use in each of the decoding layers}

\item{max.length}{Maximum length of each name}

\item{num.encode.layers}{Number of encoding layers to use}

\item{num.decode.layers}{Number of decoding layers to use}

\item{batch.size}{Batch size of model training}

\item{epochs}{Number of epochs to train for}

\item{lr}{Learning rate of model. This model is configured to use to the Adam optimizer.}

\item{validation.split}{Percentage of data to save for validation}

\item{save.dir}{Directory to save the trained encoder}

\item{reconstruct}{Whether or not show reconstructions}

\item{reconstruct.n}{How many reconstructions to show}

\item{reconstruct.display}{After many epochs to show reconstructions}

\item{earlystop}{TRUE if stopping early when validation loss is no longer decreasing,
if FALSE then train for all epochs}

\item{earlystop.patience}{Number of epochs to wait while validation loss does not
decrease before stopping training early}

\item{tensorboard}{TRUE if tensorboard metrics are to be recorded. 
Logs are recorded in the /tmp/ directory}

\item{tensorboard.runid}{Unique identifier for the run to separate tensorboard logs}

\item{verbose}{Verbosity level for training output, 0 = silence, 1 = minimal, 2 = verbose}
}
\value{
Trained encoder model
}
\description{
Trains and saves a LSTM Variational Autoencoder
}
